# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VPN Admin API
  description: Control plane for commercial VPN (AmneziaWG).
  version: 0.1.0-rc.1
paths:
...
  /api/v1/peers:
    get:
      tags:
      - peers
      summary: List Peers
      description: 'List peers (devices). peer_id = device.id, node_id = device.server_id.

        Create peer: use POST /api/v1/devices/issue (with server_id or omit for load-balanced).

        Remove peer: use POST /api/v1/devices/{id}/revoke or /reset.'
      operationId: list_peers_api_v1_peers_get
      security:
      - HTTPBearer: []
      parameters:
      - name: node_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by node (server) id
          title: Node Id
        description: Filter by node (server) id
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: active | revoked
          title: Status
        description: active | revoked
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerListOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/agent/heartbeat:
    post:
      tags:
      - agent
      summary: Agent Heartbeat
      operationId: agent_heartbeat_api_v1_agent_heartbeat_post
      parameters:
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentHeartbeatIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentAckOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
  /api/v1/agent/desired-state:
    get:
      tags:
      - agent
      summary: Agent Desired State
      operationId: agent_desired_state_api_v1_agent_desired_state_get
      parameters:
      - name: server_id
        in: query
        required: true
        schema:
          type: string
          title: Server Id
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDesiredStateOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
  /api/v1/agent/v1/status:
    get:
      tags:
      - agent
      summary: Agent V1 Status
      operationId: agent_v1_status_api_v1_agent_v1_status_get
      parameters:
      - name: server_id
        in: query
        required: true
        schema:
          type: string
          description: Server ID this agent manages
          title: Server Id
        description: Server ID this agent manages
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentV1StatusOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
  /api/v1/agent/v1/telemetry:
    get:
      tags:
      - agent
      summary: Agent V1 Telemetry
      operationId: agent_v1_telemetry_api_v1_agent_v1_telemetry_get
      parameters:
      - name: server_id
        in: query
        required: true
        schema:
          type: string
          description: Server ID this agent manages
          title: Server Id
        description: Server ID this agent manages
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentV1TelemetryOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
  /api/v1/agent/v1/peers:
    get:
      tags:
      - agent
      summary: Agent V1 Peers
      operationId: agent_v1_peers_api_v1_agent_v1_peers_get
      parameters:
      - name: server_id
        in: query
        required: true
        schema:
          type: string
          description: Server ID this agent manages
          title: Server Id
        description: Server ID this agent manages
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentV1PeerOut'
                title: Response Agent V1 Peers Api V1 Agent V1 Peers Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
  /api/v1/agent/v1/actions/execute:
    post:
      tags:
      - agent
      summary: Agent V1 Actions Execute
      operationId: agent_v1_actions_execute_api_v1_agent_v1_actions_execute_post
      parameters:
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentV1ActionExecuteIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
  /api/v1/agent/v1/actions/poll:
    get:
      tags:
      - agent
      summary: Agent V1 Actions Poll
      operationId: agent_v1_actions_poll_api_v1_agent_v1_actions_poll_get
      parameters:
      - name: server_id
        in: query
        required: true
        schema:
          type: string
          description: Server ID this agent manages
          title: Server Id
        description: Server ID this agent manages
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentV1ActionPollOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
  /api/v1/agent/v1/actions/report:
    post:
      tags:
      - agent
      summary: Agent V1 Actions Report
      operationId: agent_v1_actions_report_api_v1_agent_v1_actions_report_post
      parameters:
      - name: X-Agent-Token
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Agent-Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentV1ActionReportIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
      security:
      - AgentToken: []
      x-mtls-required: true
...
components:
  schemas:
...
    AgentAckOut:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        server_time_utc:
          type: string
          format: date-time
          title: Server Time Utc
      type: object
      required:
      - server_time_utc
      title: AgentAckOut
...
    AgentDesiredPeer:
      properties:
        public_key:
          type: string
          title: Public Key
        allowed_ips:
          type: string
          title: Allowed Ips
          default: 0.0.0.0/0, ::/0
        persistent_keepalive:
          type: integer
          title: Persistent Keepalive
          default: 25
      type: object
      required:
      - public_key
      title: AgentDesiredPeer
    AgentDesiredStateOut:
      properties:
        server_id:
          type: string
          title: Server Id
        interface_name:
          type: string
          title: Interface Name
          default: awg0
        revision:
          type: string
          title: Revision
          description: Opaque revision for idempotency/caching
        peers:
          items:
            $ref: '#/components/schemas/AgentDesiredPeer'
          type: array
          title: Peers
      type: object
      required:
      - server_id
      - revision
      - peers
      title: AgentDesiredStateOut
    AgentHeartbeatIn:
      properties:
        server_id:
          type: string
          title: Server Id
          description: Control-plane server_id this agent manages
        container_name:
          type: string
          title: Container Name
          description: Local AmneziaWG container name (amnezia-awg*)
        host_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Host Id
          description: Stable host id
        container_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Container Id
          description: Docker container id
        classification:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Classification
          description: Discovery classification
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
          description: Discovery confidence 0..1
        evidence:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Evidence
          description: Evidence strings
        interface_name:
          type: string
          title: Interface Name
          default: awg0
        public_key:
          type: string
          title: Public Key
          default: ''
        listen_port:
          type: integer
          maximum: 65535.0
          minimum: 0.0
          title: Listen Port
          default: 0
        peer_count:
          type: integer
          minimum: 0.0
          title: Peer Count
          default: 0
        total_rx_bytes:
          type: integer
          minimum: 0.0
          title: Total Rx Bytes
          default: 0
        total_tx_bytes:
          type: integer
          minimum: 0.0
          title: Total Tx Bytes
          default: 0
        health_score:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Health Score
          default: 1.0
        status:
          type: string
          title: Status
          description: healthy|degraded|unhealthy|unknown
          default: unknown
        agent_version:
          type: string
          title: Agent Version
          default: unknown
        ts_utc:
          type: string
          format: date-time
          title: Ts Utc
          description: Agent timestamp UTC
      type: object
      required:
      - server_id
      - container_name
      - ts_utc
      title: AgentHeartbeatIn
      description: Agent heartbeat payload (no secrets).
    AgentV1ActionExecuteIn:
      properties:
        type:
          type: string
          title: Type
          description: Action type e.g. sync, apply_peers
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
      type: object
      required:
      - type
      title: AgentV1ActionExecuteIn
      description: 'Stub: push-mode action request (prefer poll/report).'
    AgentV1ActionPollOut:
      properties:
        action_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Id
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        payload:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Payload
      type: object
      title: AgentV1ActionPollOut
      description: 'Pull-mode: next pending action for server_id; empty if none.'
    AgentV1ActionReportIn:
      properties:
        action_id:
          type: string
          title: Action Id
        status:
          type: string
          title: Status
          description: completed | failed
        message:
          type: string
          title: Message
          default: ''
        meta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Meta
      type: object
      required:
      - action_id
      - status
      title: AgentV1ActionReportIn
      description: Agent reports action completion or failure.
    AgentV1PeerOut:
      properties:
        public_key:
          type: string
          title: Public Key
        allowed_ips:
          type: string
          title: Allowed Ips
          default: ''
        last_handshake_age_sec:
          anyOf:
          - type: integer
          - type: 'null'
          title: Last Handshake Age Sec
        rx_bytes:
          type: integer
          title: Rx Bytes
          default: 0
        tx_bytes:
          type: integer
          title: Tx Bytes
          default: 0
      type: object
      required:
      - public_key
      title: AgentV1PeerOut
      description: Single peer in agent v1 peers list (from heartbeat or snapshot).
    AgentV1StatusOut:
      properties:
        server_id:
          type: string
          title: Server Id
        agent_version:
          type: string
          title: Agent Version
          default: unknown
        interface_name:
          type: string
          title: Interface Name
          default: awg0
        last_heartbeat_ts:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Heartbeat Ts
      type: object
      required:
      - server_id
      title: AgentV1StatusOut
      description: Agent v1 status from last heartbeat (Redis).
    AgentV1TelemetryOut:
      properties:
        server_id:
          type: string
          title: Server Id
        peer_count:
          type: integer
          title: Peer Count
          default: 0
        total_rx_bytes:
          type: integer
          title: Total Rx Bytes
          default: 0
        total_tx_bytes:
          type: integer
          title: Total Tx Bytes
          default: 0
        health_score:
          type: number
          title: Health Score
          default: 1.0
        status:
          type: string
          title: Status
          default: unknown
      type: object
      required:
      - server_id
      title: AgentV1TelemetryOut
      description: Agent v1 telemetry from last heartbeat (Redis).
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PeerListItemOut:
      properties:
        peer_id:
          type: string
          title: Peer Id
        node_id:
          type: string
          title: Node Id
        user_id:
          type: integer
          title: User Id
        subscription_id:
          type: string
          title: Subscription Id
        public_key:
          type: string
          title: Public Key
        client_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Name
        status:
          type: string
          title: Status
        issued_at:
          type: string
          format: date-time
          title: Issued At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
      type: object
      required:
      - peer_id
      - node_id
      - user_id
      - subscription_id
      - public_key
      - client_name
      - status
      - issued_at
      - revoked_at
      title: PeerListItemOut
      description: Single peer in list (device as peer).
    PeerListOut:
      properties:
        peers:
          items:
            $ref: '#/components/schemas/PeerListItemOut'
          type: array
          title: Peers
        total:
          type: integer
          title: Total
      type: object
      required:
      - peers
      - total
      title: PeerListOut
      description: Peers list response.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ErrorInfo:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
      required:
      - code
      - message
    ErrorMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
        request_id:
          type: string
      required:
      - timestamp
      - code
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          nullable: true
        error:
          $ref: '#/components/schemas/ErrorInfo'
        meta:
          $ref: '#/components/schemas/ErrorMeta'
      required:
      - success
      - data
      - error
      - meta
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Bot API key (X-API-Key header).
    AgentToken:
      type: apiKey
      in: header
      name: X-Agent-Token
      description: Node-agent shared token. mTLS required at reverse-proxy.
  responses:
    ErrorResponse:
      description: Unified error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'